Addition of Imperal Gallons for UK users
In the UK, we measure fuel consumption in imperial miles per gallon, but use litres to fill. The existing conditional statement will only allow imperal gallons if the fuelUnit == gal. This may work in most parts of the world, however for us strange brits using litres this results in iffy readings. This is a corrective fix to resolve this.